requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
PyYAML==6.0.2
pydantic==2.11.3
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
asyncpg==0.30.0
starlette==0.46.2
alembic==1.15.2
psycopg2-binary==2.9.10